Add files via upload
DHDL Ltd. (an NBFC in India) is facing challenges due to an increase in loan defaults. They aim to understand which parameters contribute to loan defaults. The company has a dataset containing information on 90,000+ clients who have taken loans from them in the past, along with loan default status. The goal is to predict whether a client will default on their loan payment or not.